Bug 564819 - API revision | conditions | stream codecs review
 side work: repackage tests in `lic.base.tests` bundle according to the
notation
